Refactoring
1. rename `mod worker` to `mod parallel_search`
2. move thread-related code into `parallel_search::Graph`
